refactor: extract run function into separate module
