+++ packages/web/src/cli/qcRulesConfigValidate.ts
@@ -1,0 +1,6 @@
+import { QCRulesConfig } from 'src/algorithms/QC/qcRulesConfig'
+
+export function qcRulesConfigValidate(config: unknown) {
+  // TODO
+  return config as QCRulesConfig
+}
